common: profiler refactor  svn merge -r 2192:2193 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/branches/yui/support_gimic
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@2303 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ common/profile.c
@@ -783,7 +783,7 @@ int profile_readint(const OEMCHAR *lpAppName, const OEMCHAR *lpKeyName, int nDef
 	{
 		nSize = min(NELEMENTS(szBuffer), pfp.cchString + 1);
 		milstr_ncpy(szBuffer, pfp.lpString, nSize);
-		return milstr_solveINT(szBuffer);
+		return (int)milstr_solveINT(szBuffer);
 	}
 }
